requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
qdrant-client==1.11.3
mistralai==1.2.4
elevenlabs==1.9.0
pydantic==2.5.0
requests==2.31.0
predicthq==1.0.0
streamlit==1.28.0